+++ CHANGES.md
@@ -19,6 +19,7 @@ Release Notes.
 * Add support `returnedObj` expression for apm-customize-enhance-plugin
 * Fix the bug that httpasyncclient-4.x-plugin  puts the dirty tracing context in the connection context
 * Compatible with the versions after dubbo-2.7.14
+* Follow protocol grammar fix `GCPhrase -> GCPhase`.
 * Support zgc gc time and count metric collect
 
 #### Documentation